requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
google-cloud-storage
xgboost
joblib
mlflow
fastapi
uvicorn

